refactor: modularize ui components into feature-based structure
- Split monolithic lib/ui/components.js (851 lines) into feature-based modules
- Organized components into 7 feature directories: welcome, status, sound, theme, version, uninstall, common
- Each feature has focused components (50-100 lines) with dedicated index.js exports
- Maintained backward compatibility through main components/index.js export portal
- Updated import paths in CLI modules and test files
- All tests passing: basic validation, interactive sounds, integration tests

Benefits:
- Clear feature boundaries for better maintainability
- Smaller, focused files easier to navigate and modify
- Better discoverability by feature name
- Scalable structure for future component additions
