+++ lib/cli/commands.js
@@ -10,7 +10,7 @@ import {
   LoadingSpinner,
   UninstallWarning,
   VersionCheckDisplay
-} from '../ui/components.js';
+} from '../ui/components/index.js';
 import { PromptManager } from '../ui/prompts.js';
 import { MESSAGES, PROMPTS } from '../ui/constants.js';
