+++ lib/cli/menus.js
@@ -15,7 +15,7 @@ import {
   InteractiveTestSoundsDisplay,
   SoundConfigState,
   SoundConfigDisplay
-} from '../ui/components.js';
+} from '../ui/components/index.js';
 import { PromptManager, MenuNavigator } from '../ui/prompts.js';
 import { 
   MESSAGES,
