+++ lib/ui/components/common/index.js
@@ -1,0 +1,7 @@
+/**
+ * Common Module
+ * Shared utility components
+ */
+
+export { LoadingSpinner } from './LoadingSpinner.js';
+export { MessageBox } from './MessageBox.js';
