+++ lib/ui/components/sound/index.js
@@ -1,0 +1,9 @@
+/**
+ * Sound Module
+ * Sound configuration and testing components
+ */
+
+export { SoundConfigState } from './SoundConfigState.js';
+export { SoundConfigDisplay } from './SoundConfigDisplay.js';
+export { TestSoundsDisplay } from './TestSoundsDisplay.js';
+export { InteractiveTestSoundsDisplay } from './InteractiveTestSoundsDisplay.js';
