+++ lib/ui/components/status/index.js
@@ -1,0 +1,7 @@
+/**
+ * Status Module
+ * System status display components
+ */
+
+export { StatusBar } from './StatusBar.js';
+export { SystemInfoDisplay } from './SystemInfoDisplay.js';
