+++ lib/ui/components/theme/index.js
@@ -1,0 +1,6 @@
+/**
+ * Theme Module
+ * Theme management components
+ */
+
+export { ThemeManagementDisplay, ThemeListDisplay } from './ThemeManagementDisplay.js';
