+++ lib/ui/components/uninstall/index.js
@@ -1,0 +1,6 @@
+/**
+ * Uninstall Module
+ * Uninstall confirmation components
+ */
+
+export { UninstallWarning } from './UninstallWarning.js';
