+++ lib/ui/components/version/index.js
@@ -1,0 +1,6 @@
+/**
+ * Version Module
+ * Version checking components
+ */
+
+export { VersionCheckDisplay } from './VersionCheckDisplay.js';
