+++ lib/ui/components/welcome/index.js
@@ -1,0 +1,6 @@
+/**
+ * Welcome Module
+ * Entry experience components
+ */
+
+export { WelcomeScreen } from './WelcomeScreen.js';
